testing/web-platform/tests/workers/Worker_ErrorEvent_lineno.htm

<!DOCTYPE html>
<html>
<head>
<title> AbstractWorker ErrorEvent.lineno </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var ErrorMessage = "Error Message";
    var FileName = './support/ErrorEvent.js';
    var description = "ErrorEvent lineno attribute represents the line number where the error occurred in the script.";
    
    var t = async_test("Test Description: " + description);
                  
    var worker = new Worker(FileName);

    worker.onerror = t.step_func(function(evt)
    {
        assert_equals(evt.lineno, 3);
        t.done();        
    });
    
    worker.postMessage(ErrorMessage);
</script>
</body>
</html>
